requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sse-starlette
pytest
pytest-asyncio
